requirements.txt
cycler
matplotlib
numpy
pandas
patsy
psutil
pyparsing
python-dateutil
pytz
sas7bdat
scipy
six
statsmodels
